introduce ReportControllerHelper
